api 调整,新增工具类 ChainWrappers
