+++ mybatis-plus-extension/src/main/java/com/baomidou/mybatisplus/extension/service/additional/query/impl/QueryChainWrapper.java
@@ -15,15 +15,7 @@
  */
 package com.baomidou.mybatisplus.extension.service.additional.query.impl;
 
-import com.baomidou.mybatisplus.core.conditions.query.Query;
-import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import com.baomidou.mybatisplus.core.metadata.TableFieldInfo;
-import com.baomidou.mybatisplus.core.toolkit.ExceptionUtils;
-import com.baomidou.mybatisplus.extension.service.additional.AbstractChainWrapper;
-import com.baomidou.mybatisplus.extension.service.additional.query.ChainQuery;
-
-import java.util.function.Predicate;
 
 /**
  * Query Chain Wrapper
@@ -31,43 +23,10 @@ import java.util.function.Predicate;
  * @author miemie
  * @since 2018-12-19
  */
-@SuppressWarnings({"serial"})
-public class QueryChainWrapper<T> extends AbstractChainWrapper<T, String, QueryChainWrapper<T>, QueryWrapper<T>>
-    implements ChainQuery<T>, Query<QueryChainWrapper<T>, T, String> {
-
-    private BaseMapper<T> baseMapper;
+@Deprecated
+public class QueryChainWrapper<T> extends com.baomidou.mybatisplus.extension.conditions.query.QueryChainWrapper<T> {
 
     public QueryChainWrapper(BaseMapper<T> baseMapper) {
-        super();
-        this.baseMapper = baseMapper;
-        super.wrapperChildren = new QueryWrapper<>();
-    }
-
-    @Override
-    public QueryChainWrapper<T> select(String... columns) {
-        wrapperChildren.select(columns);
-        return typedThis;
-    }
-
-    @Override
-    public QueryChainWrapper<T> select(Predicate<TableFieldInfo> predicate) {
-        wrapperChildren.select(predicate);
-        return typedThis;
-    }
-
-    @Override
-    public QueryChainWrapper<T> select(Class<T> entityClass, Predicate<TableFieldInfo> predicate) {
-        wrapperChildren.select(entityClass, predicate);
-        return typedThis;
-    }
-
-    @Override
-    public String getSqlSelect() {
-        throw ExceptionUtils.mpe("can not use this method for \"%s\"", "getSqlSelect");
-    }
-
-    @Override
-    public BaseMapper<T> getBaseMapper() {
-        return baseMapper;
+        super(baseMapper);
     }
 }
